+++ src/document/parse/preProcess/kill_xml.js
@@ -1,5 +1,5 @@
 const parseCitation = require('../templates/citation');
-const parseLine = require('../section/sentence').parseLine;
+const parseLine = require('../../../sentence').parseLine;
 //okay, i know you're not supposed to regex html, but...
 //https://en.wikipedia.org/wiki/Help:HTML_in_wikitext
