+++ src/section/image/Image.js
@@ -1,7 +1,7 @@
 const fetch = require('node-fetch');
 const Hashes = require('jshashes');
-const toMarkdown = require('../../../../output/markdown/image');
-const toHtml = require('../../../../output/html/image');
+const toMarkdown = require('../../output/markdown/image');
+const toHtml = require('../../output/html/image');
 const server = 'https://upload.wikimedia.org/wikipedia/commons/';
 
 const encodeTitle = function(file) {
