+++ src/section/table.js
@@ -1,6 +1,6 @@
 const helpers = require('../lib/helpers');
-const parseLine = require('./sentence/').parseLine;
-const Sentence = require('./sentence/Sentence');
+const parseLine = require('../sentence/').parseLine;
+const Sentence = require('../sentence/Sentence');
 
 const table_reg = /\{\|[\s\S]+?\|\}/g; //the largest-cities table is ~70kchars.
