+++ src/sentence/formatting.js
